+++ src/code/result/error/all_error.rs
@@ -6,11 +6,11 @@
 use super::reexports::crate_errors::*;
 use crate::impl_error;
 
-impl_error![define: ErrorNotImplemented,
+impl_error![single: ErrorNotImplemented,
     DOC_ERROR_NOT_IMPLEMENTED = "The requested functionality is not implemented.",
     self+f => write!(f, "The requested functionality is not implemented."),
 ];
-impl_error![define: ErrorNotSupported,
+impl_error![single: ErrorNotSupported,
     DOC_ERROR_NOT_SUPPORTED = "The requested functionality is not supported by this type.",
     self+f => write!(f, "The requested functionality is not supported by this type."),
 ];