+++ DOCS/CHANGELOG.md
@@ -28,6 +28,7 @@ The format is based on [Keep a Changelog], and this project adheres to
   - namespaces: `Arch`, `ByteSearch`, `Char`, `Env`, `Mem`, `Ptr`, `Str`.
   - `CompressionMode`, `EncodingMode`, `Pnm`.
   - standalone error types: `ErrorNotImplemented`, `ErrorNotSupported`, `ErrorElementNotFound`, `ErrorInvalidAxisLength`, `ErrorKeyAlreadyExists`, `ErrorMismatchedDimensions`, `ErrorMismatchedIndices`, `ErrorMismatchedLength`, `ErrorNodeEmpty`, `ErrorNodeLinkNotSet`, `ErrorNodeLinkNotUnique`, `ErrorNotEnoughElements`, `ErrorNotEnoughSpace`, `ErrorOutOfBounds`, `ErrorOverflow`, `ErrorPartiallyAdded`.
+  - composite error types: `ErrorPartialSpace`.
   - `False`, `True`, `UnitBi`, `UnitSi`.
   - `HasherPengy`.
   - `Lgc16`.